+++ src/z80asm/modlink.c
@@ -37,8 +37,9 @@ int LinkLibModule( struct libfile *library, long curmodule, char *modname );
 int SearchLibfile( struct libfile *curlib, char *modname );
 char *ReadName( FILE *file );
 void SearchLibraries( char *modname );
-void CreateBinFile( void );
-void ReadNames( char *filename, FILE *file );
+void CreateBinFile(void);
+void CreateRelocFile(void);
+void ReadNames(char *filename, FILE *file);
 void ReleaseLinkInfo( void );
 static char *CheckIfModuleWanted( FILE *file, long currentlibmodule, char *modname );
 
@@ -397,31 +398,35 @@ static void patch_exprs( ExprList *exprs )
 
 				patch_word(expr->code_pos, value); 
 
-                if ( opts.relocatable )
-				{
-                    if ( expr->type == TYPE_ADDRESS )
-                    {
-                        /* Expression contains relocatable address */
-						int offset   = get_cur_module_start(); 
+				/* Expression contains relocatable address */
+				if (expr->type == TYPE_ADDRESS) {
+
+					/* save section reloc data */
+					*(intArray_push(expr->section->reloc)) = expr->code_pos + get_cur_module_start();
+				
+					/* relocate code */
+					if (opts.relocatable)
+					{
+						int offset = get_cur_module_start() + expr->section->addr;
 						int distance = expr->code_pos + offset - curroffset;
 
-                        if ( distance >= 0 && distance <= 255 )
-                        {
-                            *relocptr++ = (Byte) distance;
-                            sizeof_reloctable++;
-                        }
-                        else
-                        {
-                            *relocptr++ = 0;
-                            *relocptr++ = distance & 0xFF;
-                            *relocptr++ = distance >> 8;
-                            sizeof_reloctable += 3;
-                        }
+						if (distance > 0 && distance < 256)	// Bugfix: when zero, need to use 3 bytes
+						{
+							*relocptr++ = (Byte)distance;
+							sizeof_reloctable++;
+						}
+						else
+						{
+							*relocptr++ = 0;
+							*relocptr++ = distance & 0xFF;
+							*relocptr++ = distance >> 8;
+							sizeof_reloctable += 3;
+						}
 
-                        totaladdr++;
+						totaladdr++;
 						curroffset = expr->code_pos + offset;
-                    }
-				}
+					}
+				}                
                 break;
 
             case RANGE_DWORD:
@@ -703,12 +708,19 @@ static int LinkModule_1(char *filename, long fptr_base, Str *section_name)
 
 				xfget_count_byte_Str(file, section_name);
 				origin = xfget_int32(file);
-
+				
 				/* load bytes to section */
 				/* BUG_0015: was reading at current position in code area, swaping order of modules */
 				section = new_section(str_data(section_name));
 				if (origin >= 0)
 					section->origin = origin;
+
+				/* if creating relocatable code, ignore origin */
+				if (opts.relocatable && origin >= 0) {
+					warn_org_ignored(filename, str_data(section_name));
+					section->origin = -1;
+					section->section_split = FALSE;
+				}
 
 				patch_file_contents(file, 0, code_size);
 			}
@@ -1002,6 +1014,46 @@ CreateBinFile( void )
         puts( "Code generation completed." );
 }
 
+
+
+void
+CreateRelocFile(void)
+{
+	STR_DEFINE(filename, FILENAME_MAX);
+	Section *section;
+	SectionHashElem *iter;
+	FILE *fp;
+	char *basename;
+	int reloc_size;
+	int i;
+
+	if (!opts.relocatable){
+		if (opts.bin_file)        /* use predined output filename from command line */
+			basename = opts.bin_file;
+		else						/* create output filename, based on project filename */
+			basename = get_first_module(NULL)->filename;
+
+		/* reloc output to filename.reloc */
+		for (section = get_first_section(&iter); section != NULL;
+			section = get_next_section(&iter))
+		{
+			str_set(filename, path_remove_ext(basename));	/* "test" */
+			if (section->name && *section->name) {			/* if section name not empty */
+				str_append_char(filename, '_');
+				str_append(filename, section->name);
+			}
+
+			fp = myfopen(get_reloc_filename(str_data(filename)), "wb");
+			if (fp) {
+				reloc_size = intArray_size(section->reloc);
+				for (i = 0; i < reloc_size; i++)
+					xfput_uint16(fp, *(intArray_item(section->reloc, i)));
+
+				myfclose(fp);
+			}
+		}
+	}
+}
 
 
 int