+++ src/z80asm/t/options.t
@@ -77,6 +77,7 @@ Usage:
     .bin = Z80 binary file
     .sym = symbols file
     .map = map file
+    .reloc = reloc file
     .def = global address definition file
     .err = error file
 
@@ -646,35 +647,7 @@ ERR
 }
 
 #------------------------------------------------------------------------------
-# -R, --relocatable
-#------------------------------------------------------------------------------
-
-# copied from z80asm.c:
-# unsigned char reloc_routine[] =
-my $reloc_routine =
-"\x08\xD9\xFD\xE5\xE1\x01\x49\x00\x09\x5E\x23\x56\xD5\x23\x4E\x23".
-"\x46\x23\xE5\x09\x44\x4D\xE3\x7E\x23\xB7\x20\x06\x5E\x23\x56\x23".
-"\x18\x03\x16\x00\x5F\xE3\x19\x5E\x23\x56\xEB\x09\xEB\x72\x2B\x73".
-"\xD1\xE3\x2B\x7C\xB5\xE3\xD5\x20\xDD\xF1\xF1\xFD\x36\x00\xC3\xFD".
-"\x71\x01\xFD\x70\x02\xD9\x08\xFD\xE9";
-
-$asm = "start: jp start";
-$bin = "\xC3\x00\x00";
-my $reloc_data = "\x01\x00\x01\x00\x01";
-
-# -R
-for my $options ('-R', '--relocatable') {
-	unlink_testfiles();
-	write_file(asm_file(), $asm);
-	t_z80asm_capture("-r0 -b $options ".asm_file(), "Relocation header is 78 bytes.\n", "", 0);
-	t_binary(read_file(bin_file(), binmode => ':raw'), $reloc_routine.$reloc_data.$bin);
-}
-
-# not -R
-unlink_testfiles();
-write_file(asm_file(), $asm);
-t_z80asm_capture("-r0 -b ".asm_file(), "", "", 0);
-t_binary(read_file(bin_file(), binmode => ':raw'), $bin);
+# -R, --relocatable - tested in reloc.t
 
 #------------------------------------------------------------------------------
 # --RCMX000