+++ src/z80asm/z80asm.c
@@ -25,6 +25,7 @@ Repository: https://github.com/pauloscustodio/z88dk-z80asm
 /* external functions */
 void Z80pass2( void );
 void CreateBinFile( void );
+void CreateRelocFile(void);
 
 
 /* local functions */
@@ -331,6 +332,7 @@ int main( int argc, char *argv[] )
 			write_def_file();
 
 		CreateBinFile();
+		CreateRelocFile();
 	}
 
 	set_error_null();